Added cages and crates and atmosphere.
Added a groupd of iron bar cages rendered using vertex3f(...) to demonstrate transparency although there is still some slight issue but overall I am happy with the effect. Also added an atmosphere that in itself rotates around the main planet, this was simply a case of rendering another slightly larger sphere at the same location as the planet and then applying a transparent atmosphere texture. Also created a group of crates which are rendered this time using a build index. Changed the ground plane to be slightly smaller as I don't want the scene to feel too sparse.

+++ GraphicsProgramming/GraphicsProgramming/Camera.h
@@ -55,7 +55,7 @@ public:
 	void setMoveSpeed(float newSpeed);
 	void setWindowWidth(int* widWidth);
 	void setWindowHeight(int* winHeight);
-
+	
 protected:
 	Input* input;
 
@@ -91,8 +91,8 @@ private:
 	float mouseYDiff;
 	int* windowWidth;
 	int* windowHeight;
-	float windowCentreX;
-	float windowCentreY;
+	int windowCentreX;
+	int windowCentreY;
 
 	bool firstRender = true;
 };
